+++ entrypoints/background/runtimeMessages.ts
@@ -11,7 +11,7 @@ import { handleAutoRefreshMessage } from "~/services/autoRefreshService"
 import { handleChannelConfigMessage } from "~/services/channelConfigStorage"
 import { handleDailyBalanceHistoryMessage } from "~/services/dailyBalanceHistory/scheduler"
 import { handleExternalCheckInMessage } from "~/services/externalCheckInService"
-import { handleLdohSiteLookupMessage } from "~/services/ldohSiteLookup/background"
+import { handleLdohSiteLookupMessage } from "~/services/integrations/ldohSiteLookup/background"
 import { handleManagedSiteModelSyncMessage } from "~/services/modelSync"
 import { handleRedemptionAssistMessage } from "~/services/redemptionAssist"
 import { handleUsageHistoryMessage } from "~/services/usageHistory/scheduler"
