+++ features/LdohSiteLookup/hooks/LdohSiteLookupContext.tsx
@@ -12,18 +12,18 @@ import {
 import {
   readFreshLdohSiteListCache,
   readLdohSiteListCache,
-} from "~/services/ldohSiteLookup/cache"
+} from "~/services/integrations/ldohSiteLookup/cache"
 import {
   buildLdohSiteLookupIndex,
   matchLdohSiteForAccount,
   type LdohSiteLookupIndex,
-} from "~/services/ldohSiteLookup/matching"
-import { requestLdohSiteLookupRefreshSites } from "~/services/ldohSiteLookup/runtime"
+} from "~/services/integrations/ldohSiteLookup/matching"
+import { requestLdohSiteLookupRefreshSites } from "~/services/integrations/ldohSiteLookup/runtime"
 import type {
   LdohSiteListCache,
   LdohSiteSummary,
-} from "~/services/ldohSiteLookup/types"
-import { buildLdohSiteSearchUrlFromUrl } from "~/services/ldohSiteLookup/url"
+} from "~/services/integrations/ldohSiteLookup/types"
+import { buildLdohSiteSearchUrlFromUrl } from "~/services/integrations/ldohSiteLookup/url"
 import { getErrorMessage } from "~/utils/error"
 import { createLogger } from "~/utils/logger"
